Added BONK script
Bonk script is a golang binary which monitors syscalls with the linux kernel audit module (created alongside SELinux) and logs it in an impactful way. If need be the script can kill malicious programs and also label IPs that are doing sketchy things. The source code is under bonk-bonk.zip/tar.gz